refactor: 시간표 표시 설정 관련 기능을 담당하는 TableDisplayRepository 신설
